restructure project files
Makefiles
