+++ golang/Makefile
@@ -1,0 +1,2 @@
+sig0namectl: sig0/*.go cmd/sig0namectl/*.go
+	go build ./cmd/sig0namectl
