--- golang/cli/go.mod
@@ -1,24 +1,0 @@
-module github.com/sig0namectl/sig0namectl
-
-go 1.21.0
-
-toolchain go1.21.7
-
-require (
-	github.com/davecgh/go-spew v1.1.1
-	github.com/miekg/dns v1.1.58
-	github.com/urfave/cli v1.22.14
-)
-
-require (
-	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
-	github.com/russross/blackfriday/v2 v2.1.0 // indirect
-	github.com/shynome/doh-client v1.1.0 // indirect
-	github.com/urfave/cli/v2 v2.27.1 // indirect
-	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
-	golang.org/x/mod v0.17.0 // indirect
-	golang.org/x/net v0.24.0 // indirect
-	golang.org/x/sync v0.7.0 // indirect
-	golang.org/x/sys v0.19.0 // indirect
-	golang.org/x/tools v0.20.0 // indirect
-)
