--- golang/golang-webasm-wip/README.md
@@ -1,4 +1,0 @@
-Links:
-
-* https://golangbot.com/webassembly-using-go/
-
