--- golang/golang-webasm-wip/go.mod
@@ -1,16 +1,0 @@
-module github.com/NetworkCommons/sig0poc1
-
-go 1.21.4
-
-require (
-	github.com/miekg/dns v1.1.58
-	github.com/shynome/doh-client v1.1.0
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	golang.org/x/mod v0.14.0 // indirect
-	golang.org/x/net v0.20.0 // indirect
-	golang.org/x/sys v0.16.0 // indirect
-	golang.org/x/tools v0.17.0 // indirect
-)
